Create volume for node_modules as well
Before this change, the volume mounting the current directory blew away
the `node_modules` directory present in the image when it was built.
Delcaring a separate volume means that the lack of a `node_modules`
directory on the host (which is preferable) will still allow the
container to start.

Solution found in this [blog post].

[blog post]: https://jdlm.info/articles/2016/03/06/lessons-building-node-app-docker.html